# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Pinecone API",
    "version": "1.0.0",
    "description": "This is the API documentation for Pinecone, a vector database service."
  },
  "servers": [{
    "url": "/"
  }],
  "paths": {
    "/describe_index_stats": {
      "post": {
        "tags": [
          "Vector Operations"
        ],
        "summary": "Get statistics about an index's contents",
        "description": "Returns statistics about the index's contents, including the vector count per namespace and the number of dimensions.",
        "operationId": "describeIndexStats",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "index_name": {
                    "type": "string",
                    "description": "The name of the index for which to get statistics",
                    "example": "my_index"
                  },
                  "project_id": {
                    "type": "string",
                    "description": "The ID of the project that the index belongs to",
                    "example": "my_project"
                  },
                  "environment": {
                    "type": "string",
                    "description": "The environment of the server to use (e.g. \"prod\" or \"staging\")",
                    "example": "prod"
                  }
                },
                "required": [
                  "index_name",
                  "project_id",
                  "environment"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "namespace_stats": {
                      "type": "object",
                      "description": "Statistics for each namespace in the index",
                      "example": {
                        "my_namespace": {
                          "vector_count": 10000,
                          "dimension": 64
                        },
                        "my_other_namespace": {
                          "vector_count": 5000,
                          "dimension": 128
                        }
                      }
                    },
                    "index_stats": {
                      "type": "object",
                      "description": "Statistics for the entire index",
                      "example": {
                        "vector_count": 15000,
                        "dimension": 64
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/query": {
      "post": {
        "tags": [
          "Vector Operations"
        ],
        "summary": "Search for similar vectors in a namespace",
        "description": "Searches a namespace using a query vector and retrieves the IDs of the most similar items, along with their similarity scores.",
        "operationId": "query",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "index_name": {
                    "type": "string",
                    "description": "The name of the index to search in",
                    "example": "my_index"
                  },
                  "namespace": {
                    "type": "string",
                    "description": "The namespace to search in",
                    "example": "my_namespace"
                  },
                  "query_vector": {
                    "type": "array",
                    "description": "The query vector to search for",
                    "example": [
                      0.1,
                      0.2,
                      0.3
                    ],
                    "items": {
                      "type": "number"
                    }
                  },
                  "top_k": {
                    "type": "integer",
                    "description": "The maximum number of results to return",
                    "example": 10
                  }
                },
                "required": [
                  "index_name",
                  "namespace",
                  "query_vector",
                  "top_k"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "results": {
                      "type": "array",
                      "description": "The list of most similar item IDs and their scores",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "description": "The ID of the item",
                            "example": "my_item_id"
                          },
                          "score": {
                            "type": "number",
                            "description": "The similarity score of the item compared to the query vector",
                            "example": 0.8
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}